Add a new wrapper for the student API
No documentation for this available, so some guessing involved.

Similar to the UCard API, but in anticipation of more apis being implemented
move some of the core connection logic into a Connection class.

The UCard API stays as is for now, but should be ported to the Connection API as well.