Implement checkConnection() for all ResourceApi subclasses
This more or less copies the checks to all other APIs.
The rooms one is special since an empty ID will not result in a 404
there, but in all rooms being returned, which isn't ideal because we
don't want to wait long for a result. Instead send a random non-existing ID,
which results in a 404.