filter validation: somewhat improve the validation of filter columns and values
For both we have no documentation of what is allowed and what not and the API
also doesn't give any errors and just ignors the filter if it contains a character
that is not allowed.

For the column names switch to an allow list for column names I have seen.

For the values make a block list of characters that I tested that don't work.
Note that "%1" leads to an internal error, so there is something else going on
with that.

This is all very horrible...