GenericApi::getResource: Handle case when the CO backend returns bogus data on unsupported field values
In case the queried field is an integer on the CO side and we would query for a non-number then the
backend would return all resources instead of an error. This led to us running out of memory in that case,
while ideally we just want to return 404 there.

Avoid this by only fetching two resources max and comparing the results resource field with the queried
value (cast to a string, because we can only query strings).

An alternative might be to add a getResourceByInteger() method and leave this all up to the caller, in case
this solution doesn't work for all cases.